requirements.txt
uvicorn
fastapi
PyJWT
PyYAML
Werkzeug < 2.0.0
alitra
azure-identity
azure-keyvault-secrets
azure-storage-blob
cryptography
dacite
injector
numpy
python-dotenv
requests
transitions
